\name{gbm2sas-package}
\alias{gbm2sas-package}
\docType{package}
\title{
Convert GBM Object Trees to SAS Code
}
\description{
Writes SAS code to get predicted values from every tree of a gbm object. 
}
\details{
\tabular{ll}{
Package: \tab gbm2sas\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-10\cr
License: GPL-3\cr
}
}
\author{
John R. Dixon  <gbm2sas@gmail.com>

Maintainer: John R Dixon <gbm2sas@gmail.com>                     
}
\keyword{ gbm     }
\keyword{ sas }
