\name{gbm2sas-package}
\alias{gbm2sas-package}
\docType{package}
\title{
Convert GBM Object Trees to SAS Code
}
\description{
Writes SAS code to get predicted values from every tree of a gbm object. 
}
\author{
John R. Dixon  <gbm2sas@gmail.com>

Maintainer: John R Dixon <gbm2sas@gmail.com>                     
}
\keyword{ gbm     }
\keyword{ sas }
